requirements.txt
streamlit
opencv-python
tensorflow
keras
numpy
pillow
matplotlib
